requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
azure-common==1.1.28
azure-core==1.33.0
azure-functions==1.23.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
datasets==3.5.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
isodate==0.7.2
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.4.3
multiprocess==0.70.16
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
propcache==0.3.1
pyarrow==19.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
soupsieve==2.7
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
